Cambios nuevos

+++ src/modules/booking/booking.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Booking } from 'src/modules/booking/entities/booking.entity';
@@ -10,17 +10,25 @@ export class BookingService {
         private readonly bookingRepository: Repository<Booking>,
     ) {}
 
+    async findByBookingNumber(booking_number: number) {
+      const numberFound = await this.bookingRepository.findOne({ where: { booking_number } });
+      return numberFound;
+    }
+
     public async findAllBookings() {
-        return await this.bookingRepository
-          .createQueryBuilder('booking')
-          .select()
-          .getMany();
+        return await this.bookingRepository.find({
+          relations: ['booking_type', 'booking_origin'],
+          where: {is_active: true},
+          order: { id_booking: 'ASC'}
+        })
       }
     
       public async findOneBooking(id_booking) {
         return await this.bookingRepository.findOne({
+          relations: ['booking_type', 'booking_origin'],
           where: {
             id_booking: id_booking,
+            is_active: true
           },
         });
       }
@@ -90,6 +98,7 @@ export class BookingService {
       }
     
       public async removeBooking(id_booking: number) {
+        if (!await this.findOneBooking(id_booking)) throw new HttpException(`Booking with id ${id_booking} does not exist`, HttpStatus.NOT_FOUND);
         try {
           await this.bookingRepository.delete({
             id_booking: id_booking,
@@ -101,6 +110,28 @@ export class BookingService {
         } catch (error) {
           return new BadRequestException(error);
         }
+    }
+
+    async archive(id_booking: number) {
+      const booking = await this.findOneBooking(id_booking);
+      if (!booking) throw new HttpException(`Booking with id ${id_booking} does not exist`, HttpStatus.NOT_FOUND);
+      try {
+        booking.is_active = false;
+        await this.bookingRepository.update(id_booking, booking);
+      } catch {
+        throw new HttpException('A problem occurred while archiving the booking', HttpStatus.NOT_FOUND);
+      }
+    }
+
+    async unarchive(id_booking: number) {
+      const booking = await this.findOneBooking(id_booking);
+      if (!booking) throw new HttpException(`Booking with id ${id_booking} does not exist`, HttpStatus.NOT_FOUND);
+      try {
+        booking.is_active = true;
+        await this.bookingRepository.update(id_booking, booking);
+      } catch {
+        throw new HttpException('A problem occurred while archiving the booking', HttpStatus.NOT_FOUND);
+      }
     }
 
 }
